+++ changelog.md
@@ -2,8 +2,6 @@
 ## TODO
 	MIDI post processing
 		arpeggio
-		after grace
-		on-off verification
 	implicit notes matching with sheet document
 		tremolo type note match
 		setTempo missing issue in Once measure layout
@@ -22,6 +20,8 @@
 
 
 *	2020.9.13
+
+	+	notation.ts: fixed note duration issue in Notation.assignMatcher.
 
 	+	notation.ts: refined note assignment for afterGrace.